requirements.txt
numpy
pytest
python-dotenv
Requests
simplemma
spacy
tensorflow
tensorflow_intel
pyyaml